requirements.txt
flask
flask_cors
pickle
pillow
matplotlib
numpy
pandas
tkinter
gunicorn
